feat(ops/terraform/deploy-nixos): make target_user_ssh_key optional
In case `target_user_ssh_key` points to an empty string, nixos-copy.sh
just doesn't set `IdentityFile=` at all.

This allows using deploy-nixos without any explicitly passed ssh keys,
but picking up whatever ssh setup the user has configured locally.

+++ ops/terraform/deploy-nixos/nixos-copy.sh
@@ -1,19 +1,23 @@
 #!/usr/bin/env bash
 #
-# Copies a NixOS system to a target host, using the provided key.
+# Copies a NixOS system to a target host, using the provided key,
+# or whatever ambient key is configured if the key is not set.
 set -ueo pipefail
-
-scratch="$(mktemp -d)"
-trap 'rm -rf -- "${scratch}"' EXIT
-
-echo -n "$DEPLOY_KEY" > $scratch/id_deploy
-chmod 0600 $scratch/id_deploy
 
 export NIX_SSHOPTS="\
     -o StrictHostKeyChecking=no\
     -o UserKnownHostsFile=/dev/null\
-    -o GlobalKnownHostsFile=/dev/null\
-    -o IdentityFile=$scratch/id_deploy"
+    -o GlobalKnownHostsFile=/dev/null"
+
+# If DEPLOY_KEY was passed, write it to $scratch/id_deploy
+if [ -n "${DEPLOY_KEY-}" ]; then
+  scratch="$(mktemp -d)"
+  trap 'rm -rf -- "${scratch}"' EXIT
+
+  echo -n "$DEPLOY_KEY" > $scratch/id_deploy
+  chmod 0600 $scratch/id_deploy
+  export NIX_SSHOPTS="$NIX_SSHOPTS -o IdentityFile=$scratch/id_deploy"
+fi
 
 nix-copy-closure \
   --to ${TARGET_USER}@${TARGET_HOST} \
